updated mermaid to version 8.4
