+++ previewer/package.json
@@ -12,11 +12,11 @@
   "license": "MIT",
   "dependencies": {
     "@fortawesome/fontawesome-free-webfonts": "^1.0.9",
-    "mermaid": "^8.2.3",
-    "react": "^16.8.6",
-    "react-dom": "^16.8.6"
+    "mermaid": "^8.4.3",
+    "react": "^16.12.0",
+    "react-dom": "^16.12.0"
   },
   "devDependencies": {
-    "parcel-bundler": "^1.12.3"
+    "parcel-bundler": "^1.12.4"
   }
 }
